requirements.txt
pyserial
pygame
socket
threading
time